fix(lint): address simple staticcheck issues
- Remove unnecessary type conversion in pow/pow_test.go
- Simplify nil slice check in utils/bytes.go (len check is sufficient)
- Convert if-else chain to switch in rpc_client/utils.go
- Change t.Error to t.Fatal for nil checks in encryptedfile_test.go

These fixes address QF1003, S1009, and SA5011 lint warnings.

+++ pow/pow_test.go
@@ -37,7 +37,7 @@ func TestPowStatus_String(t *testing.T) {
 
 func TestGetThresholdByDifficulty_Zero(t *testing.T) {
 	threshold := GetThresholdByDifficulty(big.NewInt(0))
-	expected := uint64(^uint64(0)) // Max uint64
+	expected := ^uint64(0) // Max uint64
 
 	if threshold != expected {
 		t.Errorf("GetThresholdByDifficulty(0) = %d, want %d", threshold, expected)
